Extract ListNavigationPanel
